make Redis optional

+++ app/Jobs/processXML.php
@@ -8,7 +8,6 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Redis;
-use App\Models\Provider;
 use App\Models\EPG;
 use Carbon\Carbon;
 
@@ -60,6 +59,6 @@ class processXML implements ShouldQueue
             $z->next('programme');
         }
         Storage::delete($this->file_name);
-        Redis::flushDB();
+        if(config('redis_enabled')) Redis::flushDB();
     }
 }
